+++ package.json
@@ -42,7 +42,7 @@
   },
   "scripts": {
     "build": "tsc",
-    "test": "BUNDLER_URL=https://testapi.balloondogs.team:4338 jest",
+    "test": "BUNDLER_URL=https://bundler.dev.balloondogs.network jest",
     "lint": "eslint .",
     "format": "prettier --write ."
   },
